+++ deps/miniupnpc/miniwget.c
@@ -453,7 +453,7 @@ miniwget3(const char * host,
 	/* sending the HTTP request */
 	while(sent < len)
 	{
-		n = send(s, buf+sent, len-sent, 0);
+		n = (int)send(s, buf+sent, len-sent, 0);
 		if(n < 0)
 		{
 			perror("send");
@@ -546,7 +546,7 @@ parseURL(const char * url,
 			/* "%25" is just '%' in URL encoding */
 			if(scope[0] == '2' && scope[1] == '5')
 				scope += 2;	/* skip "25" */
-			l = p2 - scope;
+			l = (int)(p2 - scope);
 			if(l >= IF_NAMESIZE)
 				l = IF_NAMESIZE - 1;
 			memcpy(tmp, scope, l);
