+++ menu/menu_animation.c
@@ -289,15 +289,15 @@ static float easing_out_in_bounce(float t, float b, float c, float d)
 static void menu_animation_ticker_generic(uint64_t idx,
       size_t max_width, size_t *offset, size_t *width)
 {
-   int ticker_period     = 2 * (*width - max_width) + 4;
+   int ticker_period     = (int)(2 * (*width - max_width) + 4);
    int phase             = idx % ticker_period;
 
    int phase_left_stop   = 2;
-   int phase_left_moving = phase_left_stop + (*width - max_width);
+   int phase_left_moving = (int)(phase_left_stop + (*width - max_width));
    int phase_right_stop  = phase_left_moving + 2;
 
    int left_offset       = phase - phase_left_stop;
-   int right_offset      = (*width - max_width) - (phase - phase_right_stop);
+   int right_offset      = (int)((*width - max_width) - (phase - phase_right_stop));
 
    if (phase < phase_left_stop)
       *offset = 0;
